Sample avgProductionTime을 double(분 단위, 소수점 허용)로 전환
- Sample: avgProductionTime을 int -> double로 변경
- ProductionLine: totalMinutes를 double로 전환, 소수점 분(초 단위 정밀도)
  까지 반영해 wall-clock 완료 판정
- Repository: avgProductionTime JSON 직렬화/역직렬화를 double로 처리
- ConsoleView: 평균 생산시간 소수점 입력 지원 및 표시 서식 조정
- MenuController: avgProductionTime 처리 타입을 double로 통일
- 생산 완료 시 재고 증가량이 실 생산량과 정확히 일치, 영속화
  사이클(저장/재로드)에서도 유지됨을 통합 테스트로 검증
- 미등록 시료는 주문 생성/생산 큐 진입이 불가함을 확인 및 회귀
  테스트 추가
- SKILL.md에 4가지 조건 반영
- 전체 테스트 230 assertions/52 test cases 통과

+++ SampleOrderSystem/controller/MenuController.h
@@ -158,7 +158,7 @@ private:
     {
         std::string id = view_.ReadSampleId();
         std::string name = view_.ReadSampleName();
-        int avgProductionTime = view_.ReadAvgProductionTime();
+        double avgProductionTime = view_.ReadAvgProductionTime();
         double yieldRate = view_.ReadYieldRate();
         int initialStock = view_.ReadInitialStock();
